requirements.txt
auto_mix_prep==0.2.0
einops==0.8.2
h5py==3.15.1
hydra-core==1.3.2
matplotlib==3.10.8
numpy==2.4.2
omegaconf==2.3.0
openai==2.16.0
opencv_python==4.11.0.86
pandas==3.0.0
Pillow==12.1.0
pytorch_lightning==2.5.0.post0
termcolor==3.3.0
torch==2.6.0
torchmetrics==1.6.2
torchvision==0.21.0
tqdm==4.67.1
transformers==4.57.6
